Fix versioning issue with Jetson Nano 2Gb
A recent commit broke the Jetson Nano 2Gb board version. In the `DISK_IMAGE_VERSION` function the version was set to `1.2.2`, but when attempting to retrieve this version in the `PLACEHOLDERS_VERSION` function, such a version didn't exist in `board_to_placeholders_version`. Updated the version to be `1.2.2` in both functions.